requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
Django==3.0.8
djangorestframework==3.11.0
importlib-metadata==1.7.0
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
wcwidth==0.2.5
zipp==3.1.0
beautifulsoup4>=4.8.0
bs4>=0.0.1
certifi>=2019.6.16
colorama>=0.4.1
lxml>=4.4.0
PySocks>=1.7.0
requests>=2.22.0
requests-futures>=1.0.0
soupsieve>=1.9.2
stem>=1.8.0
torrequest>=0.1.0
